requirements.txt
# Telegram Bot
aiogram>=3.0.0,<4.0.0

# Telegram Parsing (userbot)
telethon>=1.34.0

# AI Analysis
google-generativeai>=0.5.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Scheduling
apscheduler>=3.10.0

# Configuration
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
hypothesis>=6.100.0
aiosqlite>=0.19.0
